Move things out of types.ts.
