+++ src/client/common/logger.ts
@@ -1,10 +1,10 @@
 export {
     // aliases:
     // (temporarily for backward compatibility and to avoid churn)
+    TraceOptions as LogOptions,
     traceDecorators,
     traceError,
     traceInfo,
     traceVerbose,
     traceWarning
 } from '../logging/_trace';
-export { TraceOptions as LogOptions } from '../logging/types';
